ADMIN and SUPERADMIN can now delete a user

+++ src/resolvers/Mutation/deleteUser.js
@@ -1,14 +1,36 @@
-const deleteUser = async (parent, args, { prisma, request }, info) => {
-  const userId = getUserId(request)
+import getUserId from '../../utils/getUserId'
 
-  return prisma.mutation.deleteUser(
-    {
-      where: {
-        id: userId
-      }
-    },
-    info
-  )
+const deleteUser = async (parent, { id }, { prisma, request }, info) => {
+  // Get user ID from token (will throw an error if unauthenticated)
+  const userId = getUserId(request)
+  const user = await prisma.query.user({ where: { id: userId } })
+
+  // If id is not given, delete connected user
+  if (!id) {
+    return prisma.mutation.deleteUser(
+      {
+        where: {
+          id: userId
+        }
+      },
+      info
+    )
+  }
+
+  // If id is given and connected user is admin, delete given user
+  if (['ADMIN', 'SUPERADMIN'].indexOf(user.role) > -1) {
+    return prisma.mutation.deleteUser(
+      {
+        where: {
+          id
+        }
+      },
+      info
+    )
+  }
+
+  // If id is given and connected user is not admin, throw an error
+  throw new Error('You cannot delete a user')
 }
 
 export { deleteUser as default }
